# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /traces:
    get:
      operationId: listTraces
      tags: [traces]
      summary: List agent/LLM execution spans
      description: >-
        Recent agent/model spans pulled live from Datadog LLM Observability,
        scoped to the reporting window and optionally filtered by span kind and a
        free-text query. When the org has no LLM Obs data yet, `noData` is true
        and `spans` is empty (not an error).
      parameters:
        - $ref: "#/components/parameters/FromDate"
        - $ref: "#/components/parameters/ToDate"
        - $ref: "#/components/parameters/SpanKind"
        - $ref: "#/components/parameters/TraceQuery"
      responses:
        "200":
          description: Matching spans (possibly empty)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraceList"
  /traces/{traceId}:
    get:
      operationId: getTrace
      tags: [traces]
      summary: Single trace with its full span timeline
      description: >-
        Every span sharing the given traceId, ordered by start time, with
        wall-clock bounds and aggregate totals — the data needed to render a
        step-by-step waterfall for one agent run. `found` is false when no span
        matches the traceId in the reporting window.
      parameters:
        - name: traceId
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/FromDate"
        - $ref: "#/components/parameters/ToDate"
      responses:
        "200":
          description: The trace's spans and aggregates (possibly empty)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraceDetail"
  /traces/summary:
    get:
      operationId: getTraceSummary
      tags: [traces]
      summary: Aggregate totals for the matching spans
      description: >-
        Span count, error count, token totals, and average latency for the spans
        matching the same window and filters as the trace list.
      parameters:
        - $ref: "#/components/parameters/FromDate"
        - $ref: "#/components/parameters/ToDate"
        - $ref: "#/components/parameters/SpanKind"
        - $ref: "#/components/parameters/TraceQuery"
      responses:
        "200":
          description: Aggregate trace summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraceSummary"
components:
  parameters:
    FromDate:
      name: from
      in: query
      required: false
      description: >-
        Inclusive start of the reporting window as an ISO date (YYYY-MM-DD).
        When omitted, aggregation starts from the earliest usage event.
      schema:
        type: string
        format: date
    ToDate:
      name: to
      in: query
      required: false
      description: >-
        Inclusive end of the reporting window as an ISO date (YYYY-MM-DD).
        Events on this day are included. When omitted, aggregation runs to the
        latest usage event.
      schema:
        type: string
        format: date
    SpanKind:
      name: kind
      in: query
      required: false
      description: >-
        Restrict to a single LLM Obs span kind (e.g. llm, agent, workflow, tool,
        task, embedding, retrieval). When omitted, all kinds are returned.
      schema:
        type: string
    TraceQuery:
      name: q
      in: query
      required: false
      description: >-
        Free-text filter matched against span name, model, provider, kind, and
        ml_app.
      schema:
        type: string
  schemas:
...
    TraceSpan:
      type: object
      required:
        - spanId
        - traceId
        - name
        - kind
        - model
        - provider
        - inputTokens
        - outputTokens
        - totalTokens
        - estimatedCostUsd
        - latencyMs
        - status
        - timestamp
        - mlApp
        - tags
      properties:
        spanId:
          type: string
        traceId:
          type: string
        parentId:
          type: ["string", "null"]
        name:
          type: string
        kind:
          type: string
          description: LLM Obs span kind (llm, agent, workflow, tool, task, etc.)
        model:
          type: ["string", "null"]
        provider:
          type: ["string", "null"]
        inputTokens:
          type: integer
        outputTokens:
          type: integer
        totalTokens:
          type: integer
        estimatedCostUsd:
          type: number
          description: >-
            Datadog-estimated total cost for the span in USD (converted from
            Datadog's micro-dollar units). Independent of the per-model pricing
            used elsewhere in the app.
        latencyMs:
          type: number
          description: Span duration in milliseconds
        status:
          type: string
          description: ok or error
        timestamp:
          type: string
          description: Span start time as an ISO 8601 timestamp
        mlApp:
          type: ["string", "null"]
          description: Datadog ml_app the span was emitted under
        tags:
          type: array
          items:
            type: string
        input:
          type: ["string", "null"]
          description: Flattened span input (meta.input) as readable text
        output:
          type: ["string", "null"]
          description: Flattened span output (meta.output) as readable text
    TraceDetail:
      type: object
      required:
        - traceId
        - noData
        - found
        - startTime
        - endTime
        - durationMs
        - spans
        - spanCount
        - errorCount
        - inputTokens
        - outputTokens
        - totalTokens
        - avgLatencyMs
      properties:
        traceId:
          type: string
        noData:
          type: boolean
          description: True when Datadog has no LLM Obs data yet (empty, not an error)
        found:
          type: boolean
          description: True when at least one span matched the requested traceId
        startTime:
          type: ["string", "null"]
          description: Earliest span start in the trace as an ISO 8601 timestamp
        endTime:
          type: ["string", "null"]
          description: Latest span end in the trace as an ISO 8601 timestamp
        durationMs:
          type: number
          description: Wall-clock duration of the whole trace in milliseconds
        spans:
          type: array
          description: All spans sharing the traceId, ordered by start time
          items:
            $ref: "#/components/schemas/TraceSpan"
        spanCount:
          type: integer
        errorCount:
          type: integer
        inputTokens:
          type: integer
        outputTokens:
          type: integer
        totalTokens:
          type: integer
        avgLatencyMs:
          type: number
          description: Mean span latency in milliseconds across the trace's spans
    TraceList:
      type: object
      required:
        - noData
        - spans
      properties:
        noData:
          type: boolean
          description: True when Datadog has no LLM Obs data yet (empty, not an error)
        spans:
          type: array
          items:
            $ref: "#/components/schemas/TraceSpan"
    TraceSummary:
      type: object
      required:
        - noData
        - spanCount
        - errorCount
        - inputTokens
        - outputTokens
        - totalTokens
        - estimatedCostUsd
        - avgLatencyMs
      properties:
        noData:
          type: boolean
          description: True when Datadog has no LLM Obs data yet (empty, not an error)
        spanCount:
          type: integer
        errorCount:
          type: integer
        inputTokens:
          type: integer
        outputTokens:
          type: integer
        totalTokens:
          type: integer
        estimatedCostUsd:
          type: number
          description: >-
            Total Datadog-estimated cost in USD across matching spans.
            Independent of the per-model pricing used elsewhere in the app.
        avgLatencyMs:
          type: number
          description: Mean span latency in milliseconds across matching spans