# lib/api-zod/src/generated/types/traceDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TraceSpan } from './traceSpan';

export interface TraceDetail {
  traceId: string;
  /** True when Datadog has no LLM Obs data yet (empty, not an error) */
  noData: boolean;
  /** True when at least one span matched the requested traceId */
  found: boolean;
  /**
     * Earliest span start in the trace as an ISO 8601 timestamp
     * @nullable
     */
  startTime: string | null;
  /**
     * Latest span end in the trace as an ISO 8601 timestamp
     * @nullable
     */
  endTime: string | null;
  /** Wall-clock duration of the whole trace in milliseconds */
  durationMs: number;
  /** All spans sharing the traceId, ordered by start time */
  spans: TraceSpan[];
  spanCount: number;
  errorCount: number;
  inputTokens: number;
  outputTokens: number;
  totalTokens: number;
  /** Mean span latency in milliseconds across the trace's spans */
  avgLatencyMs: number;
}