# lib/api-zod/src/generated/types/traceSpan.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TraceSpan {
  spanId: string;
  traceId: string;
  /** @nullable */
  parentId?: string | null;
  name: string;
  /** LLM Obs span kind (llm, agent, workflow, tool, task, etc.) */
  kind: string;
  /** @nullable */
  model: string | null;
  /** @nullable */
  provider: string | null;
  inputTokens: number;
  outputTokens: number;
  totalTokens: number;
  /** Datadog-estimated total cost for the span in USD (converted from Datadog's micro-dollar units). Independent of the per-model pricing used elsewhere in the app. */
  estimatedCostUsd: number;
  /** Span duration in milliseconds */
  latencyMs: number;
  /** ok or error */
  status: string;
  /** Span start time as an ISO 8601 timestamp */
  timestamp: string;
  /**
     * Datadog ml_app the span was emitted under
     * @nullable
     */
  mlApp: string | null;
  tags: string[];
  /**
     * Flattened span input (meta.input) as readable text
     * @nullable
     */
  input?: string | null;
  /**
     * Flattened span output (meta.output) as readable text
     * @nullable
     */
  output?: string | null;
}